refactor(scoring): move BM25 keyword extraction from utils to scoring module
- Move bm25.rs from src/utils/ to src/scoring/
- Update import paths in orchestrator.rs, subagent.rs, reasoning.rs,
  and preprocessor.rs to reference scoring::bm25 instead of utils::bm25
- Create new scoring module with bm25, combine, and relevance submodules
- Remove bm25 exports from utils module
- Add scoring module declaration in lib.rs
