+++ rust/src/scoring/combine.rs
@@ -1,0 +1,9 @@
+// Copyright (c) 2026 vectorless developers
+// SPDX-License-Identifier: Apache-2.0
+
+//! Score combination strategies — weighted, cascade, and learned combinations.
+//!
+//! This module will provide strategies for combining scores from multiple
+//! scoring sources (BM25, relevance, etc.) into a final ranking score.
+//!
+//! TODO: Implement when Phase 4 (rerank/) is built.
