+++ rust/src/scoring/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2026 vectorless developers
+// SPDX-License-Identifier: Apache-2.0
+
+//! Scoring and ranking strategies.
+//!
+//! Provides unified scoring infrastructure used by agent, query, and rerank modules.
+
+pub mod bm25;
+pub mod combine;
+pub mod relevance;
+
+pub use bm25::extract_keywords;
