+++ rust/src/scoring/relevance.rs
@@ -1,0 +1,9 @@
+// Copyright (c) 2026 vectorless developers
+// SPDX-License-Identifier: Apache-2.0
+
+//! Relevance scoring — LLM-based scoring for post-retrieval reranking.
+//!
+//! This module will provide relevance scoring that combines BM25 scores with
+//! LLM-based judgments for final result ranking. Used by `rerank/` module.
+//!
+//! TODO: Implement when Phase 4 (rerank/) is built.
